+++ dash_bio/NeedlePlot.py
@@ -38,6 +38,21 @@ Keyword arguments:
 - height (number | string; default 800):
     Height of the Plot.
 
+- loading_state (dict; optional):
+    Object that holds the loading state object coming from
+    dash-renderer.
+
+    `loading_state` is a dict with keys:
+
+    - component_name (string; optional):
+        Holds the name of the component that is loading.
+
+    - is_loading (boolean; optional):
+        Determines if the component is loading or not.
+
+    - prop_name (string; optional):
+        Holds which property is loading.
+
 - margin (dict; default {t: 100, l: 40, r: 0, b: 40}):
     Margins of the plot.
 
@@ -83,12 +98,12 @@ Keyword arguments:
 - ylabel (string; optional):
     Title of the y-axis."""
     @_explicitize_args
-    def __init__(self, id=Component.UNDEFINED, mutationData=Component.UNDEFINED, margin=Component.UNDEFINED, xlabel=Component.UNDEFINED, ylabel=Component.UNDEFINED, rangeSlider=Component.UNDEFINED, needleStyle=Component.UNDEFINED, domainStyle=Component.UNDEFINED, clickData=Component.UNDEFINED, width=Component.UNDEFINED, height=Component.UNDEFINED, **kwargs):
-        self._prop_names = ['id', 'clickData', 'domainStyle', 'height', 'margin', 'mutationData', 'needleStyle', 'rangeSlider', 'width', 'xlabel', 'ylabel']
+    def __init__(self, id=Component.UNDEFINED, mutationData=Component.UNDEFINED, margin=Component.UNDEFINED, xlabel=Component.UNDEFINED, ylabel=Component.UNDEFINED, rangeSlider=Component.UNDEFINED, needleStyle=Component.UNDEFINED, domainStyle=Component.UNDEFINED, clickData=Component.UNDEFINED, width=Component.UNDEFINED, height=Component.UNDEFINED, loading_state=Component.UNDEFINED, **kwargs):
+        self._prop_names = ['id', 'clickData', 'domainStyle', 'height', 'loading_state', 'margin', 'mutationData', 'needleStyle', 'rangeSlider', 'width', 'xlabel', 'ylabel']
         self._type = 'NeedlePlot'
         self._namespace = 'dash_bio'
         self._valid_wildcard_attributes =            []
-        self.available_properties = ['id', 'clickData', 'domainStyle', 'height', 'margin', 'mutationData', 'needleStyle', 'rangeSlider', 'width', 'xlabel', 'ylabel']
+        self.available_properties = ['id', 'clickData', 'domainStyle', 'height', 'loading_state', 'margin', 'mutationData', 'needleStyle', 'rangeSlider', 'width', 'xlabel', 'ylabel']
         self.available_wildcard_properties =            []
         _explicit_args = kwargs.pop('_explicit_args')
         _locals = locals()