+++ src/lib/components/Ideogram.react.js
@@ -392,6 +392,24 @@ Ideogram.propTypes = {
      * https://eweitz.github.io/ideogram/annotations-heatmap
      * heatmaps: PropTypes.arrayOf(PropTypes.object),
      */
+
+    /**
+     * Object that holds the loading state object coming from dash-renderer
+     */
+    loading_state: PropTypes.shape({
+        /**
+         * Determines if the component is loading or not
+         */
+        is_loading: PropTypes.bool,
+        /**
+         * Holds which property is loading
+         */
+        prop_name: PropTypes.string,
+        /**
+         * Holds the name of the component that is loading
+         */
+        component_name: PropTypes.string,
+    }),
 };
 
 export const defaultProps = Ideogram.defaultProps;