+++ src/lib/components/Molecule2dViewer.react.js
@@ -76,6 +76,30 @@ Molecule2dViewer.propTypes = {
             })
         ),
     }),
+
+    /**
+     * Object that holds the loading state object coming from dash-renderer
+     */
+    loading_state: PropTypes.shape({
+        /**
+         * Determines if the component is loading or not
+         */
+        is_loading: PropTypes.bool,
+        /**
+         * Holds which property is loading
+         */
+        prop_name: PropTypes.string,
+        /**
+         * Holds the name of the component that is loading
+         */
+        component_name: PropTypes.string,
+    }),
+
+    /**
+     * The option of whether or not to allow scrolling to control
+     * the zoom.
+     */
+    scrollZoom: PropTypes.bool,
 };
 
 export const defaultProps = Molecule2dViewer.defaultProps;