+++ src/lib/components/Molecule3dViewer.js
@@ -35,6 +35,10 @@ Molecule3dViewer.defaultProps = {
         animationDuration: 0,
         fixedPath: false,
     },
+    style: {
+        height: 500,
+        width: 500,
+    },
 };
 
 Molecule3dViewer.propTypes = {
@@ -188,6 +192,21 @@ Molecule3dViewer.propTypes = {
     shapes: PropTypes.arrayOf(PropTypes.object),
 
     /**
+     * The height (in px) of the container
+     */
+    height: PropTypes.number,
+
+    /**
+     * The width (in px) of the container
+     */
+    width: PropTypes.number,
+
+    /**
+     * Generic style overrides on the plot div
+     */
+    style: PropTypes.object,
+
+    /**
      * Callback to re-render molecule viewer
      * when modelData is changed
      */
@@ -198,6 +217,24 @@ Molecule3dViewer.propTypes = {
      * when a selection is made
      */
     onChangeSelection: PropTypes.func,
+
+    /**
+     * Object that holds the loading state object coming from dash-renderer
+     */
+    loading_state: PropTypes.shape({
+        /**
+         * Determines if the component is loading or not
+         */
+        is_loading: PropTypes.bool,
+        /**
+         * Holds which property is loading
+         */
+        prop_name: PropTypes.string,
+        /**
+         * Holds the name of the component that is loading
+         */
+        component_name: PropTypes.string,
+    }),
 };
 
 export const defaultProps = Molecule3dViewer.defaultProps;