+++ src/lib/components/Speck.react.js
@@ -24,6 +24,12 @@ export default class Speck extends Component {
 Speck.defaultProps = {
     view: speckView.new(),
     data: [],
+    showLegend: false,
+    style: {
+        height: '600px',
+        width: '600px',
+        position: 'relative',
+    },
 };
 
 Speck.propTypes = {
@@ -97,6 +103,35 @@ Speck.propTypes = {
      */
 
     setProps: PropTypes.func,
+
+    /**
+     * Generic style overrides on the plot div
+     */
+    style: PropTypes.object,
+
+    /**
+     * Object that holds the loading state object coming from dash-renderer
+     */
+    loading_state: PropTypes.shape({
+        /**
+         * Determines if the component is loading or not
+         */
+        is_loading: PropTypes.bool,
+        /**
+         * Holds which property is loading
+         */
+        prop_name: PropTypes.string,
+        /**
+         * Holds the name of the component that is loading
+         */
+        component_name: PropTypes.string,
+    }),
+
+    /**
+     * The option of whether to show color legend
+     */
+
+    showLegend: PropTypes.bool,
 };
 
 export const defaultProps = Speck.defaultProps;